Refactor ConnectionRequestActivity
