+++ app/src/main/java/com/andela/helpmebuy/fragments/RequestActivityFragment.java
@@ -1,4 +1,4 @@
-package com.andela.helpmebuy.activities;
+package com.andela.helpmebuy.fragments;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
